+++ src/main/java/fr/meya/witcher_toolkit_back/model/persistent/CompetenceSpecifiquePersonnage.java
@@ -16,7 +16,7 @@ public class CompetenceSpecifiquePersonnage {
 
 	@Id
 	@GeneratedValue( strategy = GenerationType.IDENTITY)
-	private int idCompetenceSpecifiquePersonnage;
+	private long idCompetenceSpecifiquePersonnage;
 
 	private int valeurActuelle;
